added return type check
